FIRST POST (commit)
Started the basics of a new disk widget for awesome. Goal is to be able
to mount and umount drives from the widget as well as show basic
information like space and io. Right now just have enough code to
enumerate the current drives visible and load that into a table.

Also created a test script for the widget using the lua testing framwork
`busted'. Still haven't quite got the hang of mocking functions with it
though.