Mocking and testing
Added two new mock files (awful.util and awful.menu) to mock out awesome
functions rather than trying to mock the class functions for
awesome_disk. `awful.util.pread' now simply returns the value of
_G.lsblk_info which can be set inside a test and be reset after the
describe block pops off the stack.

Cleaned up some of the documentation in awesome-disk.lua and added a
general update function as well as an update_menu function (not
implemented yet) that will be used in combination with a timer to keep
the widget in sync with the file system.