Testing, is this thing on
Added all of the busted skeletal structure to awesome-disk. This
includes all of the neccessary mock stubbing for the test running to
complete the unit tests without encountering a library error. Added a
unit target to .busted that includes the necessary path information
about unit testing this project. Created a stub unit_spec file to
contain the actual unit tests.